fit with the pattern
