+++ tfjs-layers/src/engine/executor.ts
@@ -12,10 +12,11 @@
  * Executor: Evaluates SymbolicTensor based on feeds.
  */
 
-import {cast, dispose, env, memory, Tensor, util} from '@tensorflow/tfjs-core';
+import {cast, dispose, memory, Tensor, util} from '@tensorflow/tfjs-core';
 
 import {ValueError} from '../errors';
 import {Kwargs} from '../types';
+import {LruCache} from '../utils/executor_utils';
 import {toList} from '../utils/generic_utils';
 
 import {InputLayer} from './input_layer';
@@ -179,48 +180,6 @@ export class FeedDict {
     if (this.id2Mask != null) {
       dispose(this.id2Mask);
     }
-  }
-}
-
-/**
- * LruCache: A mapping from the String to T. If the number of the entries is
- * exceeding the `maxEntries`, the LruCache will delete the least recently
- * used entry.
- */
-export class LruCache<T> {
-  private cache: Map<string, T>;
-  private maxEntries: number;
-
-  constructor(maxEntries?: number) {
-    this.maxEntries = maxEntries || env().getNumber('LRU_CACHE_MAX_ENTRIES');
-    this.cache = new Map<string, T>();
-  }
-
-  /**
-   * Get the entry for the key and mark it as used recently.
-   */
-  public get(key: string): T {
-    let entry: T;
-    if (this.cache.has(key)) {
-      entry = this.cache.get(key);
-      this.cache.delete(key);
-      this.cache.set(key, entry);
-    }
-    return entry;
-  }
-
-  /**
-   * Put the entry into the cache. If the key already existed, mark the key as
-   * used recently.
-   */
-  public put(key: string, value: T) {
-    if (this.cache.has(key)) {
-      this.cache.delete(key);
-    } else if (this.cache.size >= this.maxEntries) {
-      const keyToDelete = this.cache.keys().next().value;
-      this.cache.delete(keyToDelete);
-    }
-    this.cache.set(key, value);
   }
 }
