+++ tfjs-layers/src/engine/executor_test.ts
@@ -17,7 +17,7 @@ import {dispose, memory, ones, Tensor, tensor1d, tensor2d, tensor3d, zeros} from
 import * as tfl from '../index';
 import {describeMathCPU, describeMathCPUAndGPU, expectTensorsClose} from '../utils/test_utils';
 
-import {execute, ExecutionProbe, FeedDict, getTopologicalSortAndRecipientCountsForOneFetch, LruCache} from './executor';
+import {execute, ExecutionProbe, FeedDict, getTopologicalSortAndRecipientCountsForOneFetch} from './executor';
 
 // tslint:enable
 
@@ -240,23 +240,5 @@ describeMathCPUAndGPU('Executor', () => {
           outputs[1], tensor2d([10, 10, 10, 10, 10, 10], [2, 3]));
       expect(callCounter).toEqual(1);
     });
-  });
-});
-
-describeMathCPU('LruCache', () => {
-  const maxEntries = 3;
-  const cache = new LruCache<number>(maxEntries);
-
-  it('Delete the leaset recent used entry when exceeding the size', () => {
-    cache.put('1', 1);  // Caching [1]
-    cache.put('2', 2);  // Caching [1, 2]
-    cache.put('3', 3);  // Caching [1, 2, 3]
-    cache.get('1');     // Caching [2, 3, 1]
-    cache.put('4', 4);  // Caching [3, 1, 4]
-
-    expect(cache.get('2')).toBeUndefined();
-    expect(cache.get('1')).toBe(1);
-    expect(cache.get('3')).toBe(3);
-    expect(cache.get('4')).toBe(4);
   });
 });
